implemented internal parser of command-line options
As some systems have lack of getopt.h and do not support long options,
a simple internal parser of command-line options has been implemented.

Also the new parser brings an ability to distribute options
between actions. So now if an option makes no sense for an action
specified, an error message will be printed.

Also it made the program able to build successfully on Solaris.