Fix bug in ID3v2 COMM frame unpacker
When a COMM frame has an encoding requiring BOM, both description
and text fields start with BOM. So they cannot be converted to
another encoding as a continuous string, i.e. they must be converted
separately.

E.g.

$ printf "\xFF\xFE\x20\x00\x21\x00\x00\x00\xFF\xFE\x20\x00\x00\x00"
  | iconv -f UTF16 -t UTF-8 | hexdump -C

  00000000  20 21 00 ef bb bf 20 00                      | !.... .|
  00000008

Here is a garbage (ef bb bf) as a result of conversion of the second
BOM that is interpreted as a ZWNBSP.

As a side effect this commit introduces the u16_char type,
the U16_CHAR macro and the u16_memchr() function.